Auto estimation of MTU to be advertised by DHCP
Cleaning up of the interface MTU code:
1. Set Min/Max range for DHCP interface MTU and make it an integer in
   the user facing portion
2. Add a interface scan cache in DatapathController maintaining the
   list of interface descriptions from the most recent scan
   --- if that cache is not empty, the request for list of {intf
       desc, tunnel type} would simply get interface data from cache
   --- if that cache is empty, DatapathController would trigger an
       interface scan request
3. The aforementioned request message no longer assume there is only
   one tunnel type for a particular interface

+++ midonet-client/src/main/java/com/midokura/midonet/client/dto/DtoDhcpSubnet.java
@@ -16,7 +16,7 @@ public class DtoDhcpSubnet {
     private String defaultGateway;
     private String serverAddr;
     private String dnsServerAddr;
-    private short interfaceMTU;
+    private int interfaceMTU;
     private List<DtoDhcpOption121> opt121Routes;
     private URI hosts;
     private URI uri;
@@ -65,11 +65,11 @@ public class DtoDhcpSubnet {
         this.dnsServerAddr = dnsServerAddr;
     }
 
-    public short getInterfaceMTU() {
+    public int getInterfaceMTU() {
         return interfaceMTU;
     }
 
-    public void setInterfaceMTU(short interfaceMTU) {
+    public void setInterfaceMTU(int interfaceMTU) {
         this.interfaceMTU = interfaceMTU;
     }
